Move `reaction_models` into `models`
