+++ src/paper/paper_upload_tasks.py
@@ -596,7 +596,7 @@ def celery_create_paper(self, celery_data):
         uploaded_by = paper_submission.uploaded_by
 
         if uploaded_by:
-            from discussion.reaction_models import Vote
+            from discussion.models import Vote
 
             Vote.objects.create(
                 content_type=get_content_type_for_model(paper),
