+++ src/paper/views/paper_views.py
@@ -27,7 +27,7 @@ from rest_framework.permissions import (
 from rest_framework.response import Response
 
 from analytics.amplitude import track_event
-from discussion.reaction_models import Vote
+from discussion.models import Vote
 from discussion.reaction_serializers import VoteSerializer
 from discussion.reaction_views import ReactionViewActionMixin
 from hub.permissions import IsModerator
