+++ src/reputation/tests/test_bounties.py
@@ -7,7 +7,7 @@ from django.db.models import Sum
 from django.utils import timezone
 from rest_framework.test import APITestCase
 
-from discussion.reaction_models import Vote
+from discussion.models import Vote
 from hub.models import Hub
 from hub.tests.helpers import create_hub
 from paper.tests.helpers import create_paper
@@ -16,8 +16,9 @@ from reputation.distributor import Distributor
 from reputation.models import Bounty, BountyFee, BountySolution, Distribution
 from researchhub_comment.tests.helpers import create_rh_comment
 from user.models import User
-from user.tests.helpers import create_moderator, create_random_default_user, create_user
 from user.related_models.user_model import FOUNDATION_REVENUE_EMAIL
+from user.tests.helpers import create_moderator, create_random_default_user, create_user
+
 
 class BountyViewTests(APITestCase):
     def setUp(self):
@@ -1417,7 +1418,9 @@ class BountyViewTests(APITestCase):
         self.assertNotIn(unused_hub.id, hub_ids)
 
     def test_bounty_dao_fee_goes_to_community_revenue_account(self):
-        community_revenue_user, _ = User.objects.get_or_create(email=FOUNDATION_REVENUE_EMAIL)
+        community_revenue_user, _ = User.objects.get_or_create(
+            email=FOUNDATION_REVENUE_EMAIL
+        )
         self.client.force_authenticate(self.user)
         response = self.client.post(
             "/api/bounty/",
@@ -1428,5 +1431,7 @@ class BountyViewTests(APITestCase):
             },
         )
         self.assertEqual(response.status_code, 201)
-        dao_fee_distribution = Distribution.objects.filter(distribution_type="BOUNTY_DAO_FEE").latest("created_date")
+        dao_fee_distribution = Distribution.objects.filter(
+            distribution_type="BOUNTY_DAO_FEE"
+        ).latest("created_date")
         self.assertEqual(dao_fee_distribution.recipient, community_revenue_user)
