+++ src/researchhub_document/views/researchhub_unified_document_views.py
@@ -11,7 +11,7 @@ from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
-from discussion.reaction_models import Vote
+from discussion.models import Vote
 from discussion.reaction_serializers import VoteSerializer
 from paper.models import Paper
 from paper.utils import get_cache_key
